requirements.txt
streamlit
pandas
json
plotly.express